+++ pkg/diff/compare.go
@@ -1,8 +1,10 @@
 package diff
 
 import (
+	//	log "github.com/Sirupsen/logrus"
 	"github.com/kylelemons/godebug/pretty"
 	"github.com/pearsontechnology/environment-operator/pkg/bitesize"
+	"k8s.io/client-go/pkg/api/resource"
 )
 
 // Compare creates a changeMap for the diff between environment configs and returns a boolean if changes were detected
@@ -11,8 +13,8 @@ func Compare(config1, config2 bitesize.Environment) bool {
 
 	newChangeMap()
 
-	c1 := config1
-	c2 := config2
+	c1 := config1 //New Config
+	c2 := config2 //Existing Config
 
 	// Following fields are ignored for diff purposes
 	c1.Tests = []bitesize.Test{}
@@ -49,7 +51,8 @@ func Compare(config1, config2 bitesize.Environment) bool {
 
 // Can't think of a better word
 func alignServices(src, dest *bitesize.Service) {
-	//log.Debugf("Aligning Service %s", src.Name)
+	//Note: src=new config    dest=existing config
+
 	// Copy version from dest if source version is empty
 	if src.Version == "" {
 		src.Version = dest.Version
@@ -61,6 +64,37 @@ func alignServices(src, dest *bitesize.Service) {
 
 	// Copy status from dest (status is only stored in the cluster)
 	src.Status = dest.Status
+
+	//If its a TPR type service, sync up the Limits since they aren't appied to the k8s resource
+	if src.Type != "" {
+		src.Limits.Memory = dest.Limits.Memory
+		src.Limits.CPU = dest.Limits.CPU
+
+	}
+
+	//Sync up Requests in the case where different units are present, but they represent equivalent quantities
+	destmemreq, _ := resource.ParseQuantity(dest.Requests.Memory)
+	srcmemreq, _ := resource.ParseQuantity(src.Requests.Memory)
+	destcpureq, _ := resource.ParseQuantity(dest.Requests.CPU)
+	srccpureq, _ := resource.ParseQuantity(src.Requests.CPU)
+	if destmemreq.Cmp(srcmemreq) == 0 {
+		src.Requests.Memory = dest.Requests.Memory
+	}
+	if destcpureq.Cmp(srccpureq) == 0 {
+		src.Requests.CPU = dest.Requests.CPU
+	}
+
+	//Sync up Limits in the case where different units are present, but they represent equivalent quantities
+	destmemlim, _ := resource.ParseQuantity(dest.Limits.Memory)
+	srcmemlim, _ := resource.ParseQuantity(src.Limits.Memory)
+	destcpulim, _ := resource.ParseQuantity(dest.Limits.CPU)
+	srccpulim, _ := resource.ParseQuantity(src.Limits.CPU)
+	if destmemlim.Cmp(srcmemlim) == 0 {
+		src.Limits.Memory = dest.Limits.Memory
+	}
+	if destcpulim.Cmp(srccpulim) == 0 {
+		src.Limits.CPU = dest.Limits.CPU
+	}
 
 	// Override source replicas with dest replicas if HPA is active
 	if dest.HPA.MinReplicas != 0 {