+++ pkg/translator/kubemapper.go
@@ -605,45 +605,39 @@ func getAccessModesFromString(modes string) []v1.PersistentVolumeAccessMode {
 	return accessModes
 }
 
-func (w *KubeMapper) getLimits() v1.ResourceList {
-
-	cpu, err := resource.ParseQuantity(w.BiteService.Limits.CPU)
-	if err != nil {
-		cpu, _ = resource.ParseQuantity(config.Env.LimitDefaultCPU)
-	}
-	mem, err := resource.ParseQuantity(w.BiteService.Limits.Memory)
-	if err != nil {
-		mem, _ = resource.ParseQuantity(config.Env.LimitDefaultMemory)
-	}
-
-	return v1.ResourceList{
-		"cpu":    cpu,
-		"memory": mem,
-	}
-}
-
 func (w *KubeMapper) resources() (v1.ResourceRequirements, error) {
 	//Environment Operator allows for Guaranteed and Burstable QoS Classes as limits are always assigned to containers
-	cpuRequest, memoryError := resource.ParseQuantity(w.BiteService.Requests.CPU)
-	memoryRequest, cpuError := resource.ParseQuantity(w.BiteService.Requests.Memory)
+	cpuRequest, memoryRequestError := resource.ParseQuantity(w.BiteService.Requests.CPU)
+	memoryRequest, cpuRequestError := resource.ParseQuantity(w.BiteService.Requests.Memory)
+	cpuLimit, _ := resource.ParseQuantity(w.BiteService.Limits.CPU)
+	memoryLimit, _ := resource.ParseQuantity(w.BiteService.Limits.Memory)
 
-	if cpuError != nil && memoryError != nil { //If no CPU or Memory Request provided, default to limits for Guaranteed QoS
+	if cpuRequestError != nil && memoryRequestError != nil { //If no CPU or Memory Request provided, default to limits for Guaranteed QoS
 		return v1.ResourceRequirements{
-			Limits: w.getLimits(),
+			Limits: v1.ResourceList{
+				"cpu":    cpuLimit,
+				"memory": memoryLimit,
+			},
 		}, nil
-	} else if cpuError != nil && memoryError == nil {
+	} else if cpuRequestError != nil && memoryRequestError == nil {
 		return v1.ResourceRequirements{
 			Requests: v1.ResourceList{
 				"memory": memoryRequest,
 			},
-			Limits: w.getLimits(),
+			Limits: v1.ResourceList{
+				"cpu":    cpuLimit,
+				"memory": memoryLimit,
+			},
 		}, nil
-	} else if cpuError == nil && memoryError != nil {
+	} else if cpuRequestError == nil && memoryRequestError != nil {
 		return v1.ResourceRequirements{
 			Requests: v1.ResourceList{
 				"cpu": cpuRequest,
 			},
-			Limits: w.getLimits(),
+			Limits: v1.ResourceList{
+				"cpu":    cpuLimit,
+				"memory": memoryLimit,
+			},
 		}, nil
 	} else {
 		return v1.ResourceRequirements{
@@ -651,7 +645,10 @@ func (w *KubeMapper) resources() (v1.ResourceRequirements, error) {
 				"cpu":    cpuRequest,
 				"memory": memoryRequest,
 			},
-			Limits: w.getLimits(),
+			Limits: v1.ResourceList{
+				"cpu":    cpuLimit,
+				"memory": memoryLimit,
+			},
 		}, nil
 
 	}