+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/). Additionally,
 
 *  Added mongo support. Environment operator can now stand up a mongodb statefulset if specified in environments.bitesize. [[BITE-1632](https://agile-jira.pearson.com/browse/BITE-1632)]
 *  Enabled Guaranteed Quality of Service. Environment operator will now deploy containers with requests=limits when a request is specified within the manifest (environments.bitesize) for a service. [[BITE-1713](https://agile-jira.pearson.com/browse/BITE-1713))]
+*  Cleaned up documentation and added a Quick Start Guide. [[BITE-1788](https://agile-jira.pearson.com/browse/BITE-1788))]
 
 ### **[0.0.7] - 2017-09-25 [RELEASED]**